+++ client/index.html
@@ -1,5 +1,5 @@
 <!doctype html>
-<html lang="en" version="2.8">
+<html lang="en" version="2.9">
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/threadup-logo.svg" />
