+++ client/src/pages/Register.tsx
@@ -1,4 +1,4 @@
-import { RegisterForm } from '@/components/register-form';
+import { RegisterForm } from '@/components/RegisterForm';
 
 export interface RegisterProps { }
 export const Register = ({ }: RegisterProps) => {
