+++ client/src/pages/index.ts
@@ -1,5 +1,4 @@
 export { Home } from './Home';
 export { Login } from './Login';
 export { Register } from './Register';
-export { Dashboard } from './Dashboard';
 export { Profile } from './Profile';
