+++ client/src/stores/posts/types.ts
