+++ client/src/utils/getAuthHeaders.ts
@@ -1,0 +1,7 @@
+import { useAuthStore } from "@/stores/auth/useAuthStore";
+
+export const getAuthHeaders = async () => {
+  const token = useAuthStore.getState().user?.token;
+  const config = { headers: { Authorization: `Bearer ${token}` } };
+  return config;
+};
