+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "threadup",
-  "version": "2.8",
+  "version": "2.9",
   "license": "UNLICENSED",
   "author": "Romel Custodio",
   "scripts": {
