+++ server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "threadup-server",
-  "version": "2.8",
+  "version": "2.9",
   "description": "",
   "main": "index.js",
   "scripts": {
