+++ server/utils/tokens.js
@@ -2,7 +2,7 @@ const jwt = require('jsonwebtoken')
 
 // Access token: short-lived (e.g. 15m) → used for API requests.
 const generateAccessToken = (id) =>
-  jwt.sign({ id }, process.env.JWT_SECRET, { expiresIn: '15m' })
+  jwt.sign({ id }, process.env.JWT_SECRET, { expiresIn: '30d' })
 
 //Refresh token: long-lived (e.g. 7d or 30d) → used to request a new access token when the old one expires.
 const generateRefreshToken = (id) =>
